+++ src/preset/cardinal.rs
@@ -1,4 +1,8 @@
-use bevy::{ecs::spawn::SpawnableList, prelude::*};
+use bevy::{
+    ecs::spawn::SpawnableList,
+    prelude::*,
+    ptr::{MovingPtr, move_as_ptr},
+};
 
 use crate::prelude::*;
 
@@ -66,18 +70,24 @@ impl Cardinal<Binding, Binding, Binding, Binding> {
 }
 
 impl<N: Bundle, E: Bundle, S: Bundle, W: Bundle> SpawnableList<BindingOf> for Cardinal<N, E, S, W> {
-    fn spawn(self, world: &mut World, entity: Entity) {
+    fn spawn(this: MovingPtr<'_, Self>, world: &mut World, entity: Entity) {
+        let cardinal = this.read();
         let x = Bidirectional {
-            positive: self.east,
-            negative: self.west,
+            positive: cardinal.east,
+            negative: cardinal.west,
         };
-        x.spawn(world, entity);
+
+        move_as_ptr!(x);
+        SpawnableList::spawn(x, world, entity);
 
         let y = Bidirectional {
-            positive: self.north,
-            negative: self.south,
-        };
-        y.with(SwizzleAxis::YXZ).spawn(world, entity);
+            positive: cardinal.north,
+            negative: cardinal.south,
+        }
+        .with(SwizzleAxis::YXZ);
+
+        move_as_ptr!(y);
+        SpawnableList::spawn(y, world, entity);
     }
 
     fn size_hint(&self) -> usize {