+++ tests/state_and_value_merge.rs
@@ -1,5 +1,5 @@
 use bevy::{ecs::spawn::SpawnWith, input::InputPlugin, prelude::*};
-use bevy_enhanced_input::prelude::*;
+use bevy_enhanced_input::prelude::{Release, *};
 use test_log::test;
 
 #[test]